modified

+++ Final/Find_the_Missing_Number.c
@@ -9,8 +9,8 @@ int main()
     {
         long long int m;
         int a, b, c;
-        scanf("%d %d %d %d",&m, &a, &b, &c);
-        int mul = a*b*c;
+        scanf("%lld %d %d %d",&m, &a, &b, &c);
+        long long int mul = a*b*c;
 
         if(m == 0){
             printf("0");
